# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RentMitra API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/google:
    post:
      operationId: googleAuth
      tags: [auth]
      summary: Authenticate with Google OAuth token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleAuthInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── USERS ─────────────────────────────────────────────────────────────────
...
  /admin/listings:
    get:
      operationId: adminGetListings
      tags: [admin]
      summary: Get all listings (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, expired, all]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingPage"
  /admin/listings/{id}/approve:
    patch:
      operationId: approveListing
      tags: [admin]
      summary: Approve a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /admin/listings/{id}/reject:
    patch:
      operationId: rejectListing
      tags: [admin]
      summary: Reject a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectListingInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /admin/listings/{id}/feature:
    patch:
      operationId: featureListing
      tags: [admin]
      summary: Toggle featured status of a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureListingInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Get all users (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
  /admin/users/{userId}/verify:
    patch:
      operationId: verifyUser
      tags: [admin]
      summary: Verify a user account (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── AUTH ───────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 8
        phone:
          type: ["string", "null"]
        userType:
          type: string
          enum: [individual, business]
          default: individual
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    GoogleAuthInput:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
    VerifyEmailInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── USERS ──────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, userType, isVerified, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        profilePhoto:
          type: ["string", "null"]
        userType:
          type: string
          enum: [individual, business, admin]
        isVerified:
          type: boolean
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PublicUser:
      type: object
      required: [id, name, userType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        profilePhoto:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        userType:
          type: string
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        listingCount:
          type: integer
        businessProfile:
          oneOf:
            - $ref: "#/components/schemas/BusinessProfile"
            - type: "null"
...
    UserPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── LISTINGS ───────────────────────────────────────────────────────────
    RentalPrice:
      type: object
      properties:
        daily:
          type: ["number", "null"]
        weekly:
          type: ["number", "null"]
        monthly:
          type: ["number", "null"]
    Listing:
      type: object
      required: [id, ownerId, title, category, status, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        brand:
          type: ["string", "null"]
        condition:
          type: string
          enum: [new, like_new, good, fair, poor]
        rentalPrice:
          $ref: "#/components/schemas/RentalPrice"
        city:
          type: string
        state:
          type: string
        pincode:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        thumbnails:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        isFeatured:
          type: boolean
        rejectionReason:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        owner:
          oneOf:
            - $ref: "#/components/schemas/PublicUser"
            - type: "null"
...
    ListingPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
...
    BusinessProfile:
      type: object
      required: [id, userId, businessName]
      properties:
        id:
          type: integer
        userId:
          type: integer
        businessName:
          type: string
        logo:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        gstNumber:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    RejectListingInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    FeatureListingInput:
      type: object
      required: [featured]
      properties:
        featured:
          type: boolean
...