# lib/api-zod/src/generated/types/rentalPrice.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RentMitra API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RentalPrice {
  /** @nullable */
  daily?: number | null;
  /** @nullable */
  weekly?: number | null;
  /** @nullable */
  monthly?: number | null;
}