# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RentMitra API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserUserType } from './userUserType';

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  profilePhoto?: string | null;
  userType: UserUserType;
  isVerified: boolean;
  emailVerified?: boolean;
  createdAt: Date;
}